@@ -28,31 +28,32 @@
     "author": "Vitaly Yerofeyevsky",
     "license": "MPL-2.0",
     "dependencies": {
-        "autoprefixer": "^10.4.15",
+        "@jest/globals": "^29.7.0",
+        "autoprefixer": "^10.4.16",
         "chokidar": "^3.5.3",
         "cross-env": "^7.0.3",
-        "esbuild": "^0.19.2",
+        "esbuild": "^0.19.11",
         "eslint-config-airbnb": "^19.0.4",
-        "fs-extra": "^11.1.1",
+        "fs-extra": "^11.2.0",
         "kolorist": "^1.8.0",
         "qrcode": "^1.5.3",
-        "tailwindcss": "^3.3.3",
-        "web-ext": "^7.6.2",
+        "tailwindcss": "^3.4.1",
+        "web-ext": "^7.10.0",
         "webextension-polyfill": "^0.10.0"
     },
     "devDependencies": {
-        "@types/qrcode": "^1.5.1",
-        "@types/webextension-polyfill": "^0.10.1",
-        "eslint": "^8.47.0",
-        "eslint-config-prettier": "^9.0.0",
-        "eslint-plugin-import": "^2.28.1",
-        "eslint-plugin-prettier": "^5.0.0",
-        "jest": "^29.6.2",
-        "jest-environment-jsdom": "^29.6.2",
+        "@types/qrcode": "^1.5.5",
+        "@types/webextension-polyfill": "^0.10.7",
+        "eslint": "^8.56.0",
+        "eslint-config-prettier": "^9.1.0",
+        "eslint-plugin-import": "^2.29.1",
+        "eslint-plugin-prettier": "^5.1.3",
+        "jest": "^29.7.0",
+        "jest-environment-jsdom": "^29.7.0",
         "npm-run-all": "^4.1.5",
-        "prettier": "^3.0.2",
-        "rimraf": "^5.0.1",
-        "tsx": "^3.12.7"
+        "prettier": "^3.2.2",
+        "rimraf": "^5.0.5",
+        "tsx": "^4.7.0"
     },
     "jest": {
         "moduleFileExtensions": [
